Remove `Typography` component (#426)
