+++ .changeset/thin-pianos-shave.md
@@ -1,0 +1,7 @@
+---
+'@utilitywarehouse/web-ui': major
+---
+
+Remove deprecated `Typography` component. Consumers should use the `Text` &
+`Heading` components instead. Please see the [migration documentation](https://uw-web-ui.vercel.app/?path=/docs/web-ui-guides-migration-v1--documentation) for further
+details.
