+++ packages/web-ui/src/Link/Link.stories.tsx
@@ -5,7 +5,6 @@ import { Text, TextProps } from '../Text';
 import { Stack } from '../Stack';
 import { Box } from '../Box';
 import { colorsCommon } from '@utilitywarehouse/colour-system';
-import { headingVariantMapping, textVariantMapping } from '../Typography/Typography';
 import { ThemeProvider } from '../ThemeProvider';
 
 const meta: Meta<typeof Link> = {
@@ -23,8 +22,8 @@ const meta: Meta<typeof Link> = {
 export default meta;
 type Story = StoryObj<typeof Link>;
 
-const textVariants = Object.keys(textVariantMapping);
-const variants = [...Object.keys(headingVariantMapping), ...textVariants];
+const textVariants = ['subtitle', 'body', 'legalNote', 'caption'] as const;
+const variants = ['displayHeading', 'h1', 'h2', 'h3', 'h4', ...textVariants] as const;
 
 export const Workshop: Story = {
   render: args => {
