--- packages/web-ui/src/Typography/Typography.docs.mdx
@@ -1,15 +1,0 @@
-import { Meta, Canvas, ArgTypes, Description } from '@storybook/blocks';
-import Typography from './Typography.tsx';
-import * as Stories from './Typography.stories';
-import { DocsHeader } from '../storybook-utils';
-
-<Meta of={Stories} />
-
-<DocsHeader componentName="Typography" stories={Stories} themeProviderRequired />
-
-## Props
-
-All the [MUI system props](https://mui.com/system/properties/) are available on
-this component.
-
-<ArgTypes of={Stories} />
