--- packages/web-ui/src/Typography/index.ts
@@ -1,8 +1,0 @@
-export { Typography } from './Typography';
-export { legacyTypographyThemeOverrides, baseTypographyTheme } from './Typography.theme';
-export type {
-  TypographyProps,
-  DefaultLegacyTypographyComponent,
-  LegacyTypographyOwnProps,
-  LegacyTypographyTypeMap,
-} from './Typography';
