+++ packages/web-ui/src/lab/TextLink/TextLink.props.ts
@@ -1,7 +1,7 @@
-import { TypographyProps } from '../../Typography';
+import { TextProps } from '../../Text';
 
 export interface TextLinkProps
-  extends Pick<TypographyProps, 'textTransform'>,
+  extends Pick<TextProps, 'textTransform'>,
     React.ComponentPropsWithoutRef<'a'> {
   /**
    * Change the default rendered element for the one passed as a child, merging their props and behavior.
