+++ packages/web-ui/src/theme/theme.ts
@@ -1,5 +1,5 @@
 import { createTheme, Theme as MuiTheme } from '@mui/material/styles';
-import { baseTypographyTheme, legacyTypographyThemeOverrides } from '../Typography';
+import { baseTypographyTheme } from './Typography.theme';
 import { htmlFontSize, pxToRem, spacing } from '../utils';
 import { breakpoints, fonts, fontWeights } from '../tokens';
 import { colors, colorsCommon } from '@utilitywarehouse/colour-system';
@@ -7,11 +7,6 @@ import { colors, colorsCommon } from '@utilitywarehouse/colour-system';
 const theme: MuiTheme = createTheme({
   breakpoints: { values: breakpoints },
   spacing,
-  components: {
-    MuiTypography: {
-      styleOverrides: legacyTypographyThemeOverrides,
-    },
-  },
 });
 
 theme.typography = {
